requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
django-filter==23.3
pillow==11.3.0
psycopg2-binary==2.9.7
drf-yasg==1.21.7
pytest==7.4.3
pytest-django==4.5.2
coverage==7.3.2
factory-boy==3.3.0
whitenoise==6.6.0
gunicorn==21.2.0